Added commands to create clothing and shoes

+++ eShop.Domain/DTOs/ShoesDTO.cs
@@ -5,7 +5,7 @@ namespace eShop.Domain.DTOs
     public class ShoesDTO : ProductDTO
     {
         public ShoesDTO() => this.ProductType = ProductType.Shoes;
-        public byte Size { get; set; }
+        public int Size { get; set; }
         public Colors Color { get; set; }
         public Audience Audience { get; set; }
     }
