+++ eShop.Domain/Entities/Clothing.cs
@@ -5,7 +5,7 @@ namespace eShop.Domain.Entities
     public class Clothing : Product
     {
         public Clothing() => this.ProductType = ProductType.Clothing;
-        public byte Size { get; set; }
+        public int Size { get; set; }
         public Colors Color { get; set; }
         public Audience Audience { get; set; }
     }
