+++ eShop.Domain/Entities/Shoes.cs
@@ -5,7 +5,7 @@ namespace eShop.Domain.Entities
     public class Shoes : Product
     {
         public Shoes() => this.ProductType = ProductType.Shoes;
-        public byte Size { get; set; }
+        public int Size { get; set; }
         public Colors Color { get; set; }
         public Audience Audience { get; set; }
     }
